sctp_auth_get_chklist_size returns a size_t instead of an int.

+++ trunk/KERN/netinet/sctp_auth.c
@@ -157,7 +157,7 @@ sctp_auth_delete_chunk(uint8_t chunk, sctp_auth_chklist_t *list)
 	return (0);
 }
 
-inline int
+inline size_t
 sctp_auth_get_chklist_size(const sctp_auth_chklist_t *list)
 {
 	if (list == NULL)
